requirements.txt
pygame
neat-python

